changed: PM3 TV Window to new look and feel fixed: Wrong recording size was given by VTPTransceiver fixed: Play of Recordings during EPG load (temporary fix)
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/pvr-testing@21186 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ XBMC/xbmc/pvrclients/vdr/pvrclient-vdr.cpp
@@ -3668,7 +3668,7 @@ int PVRClientVDR::ReadRecordedStream(BYTE* buf, int buf_size)
   vector<string> lines;
   int            code;
   char           buffer[1024];
-  unsigned int   amountReceived;
+  unsigned long  amountReceived;
 
   if (!m_transceiver->IsOpen() || m_socket_video == INVALID_SOCKET)
     return 0;
